Complete Project

+++ migrations/20240722123004_payment.sql
@@ -1,5 +1,5 @@
 CREATE TABLE payment
 (
     payment_id UUID PRIMARY KEY,
-    amount     DECIMAL(10, 2) NOT NULL
+    amount     INTEGER NOT NULL
 );
